Refactor `upgrade-provider`
